Refactor telemetry validation out in `pkg/validation` to avoid circular imports.
